Move general utilities from `codegen/src/html.rs` to `codegen/src/main.rs`
